backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-appconf==1.0.5
django-celery-email==3.0.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.13.0
flower==2.0.1
graphviz==0.20.1
gunicorn==21.2.0
humanize==4.8.0
iamport-rest-client==0.9.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.9
